+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Renamed L-BTC to LBTC
 - Improved notifications translations
 - Persist wallet list across app reinstallation
+- Reduce login time
 
 ## [4.1.7] - 2025-02-21